Adjust visibility of types, rename constructor
